astra_server/requirements.txt
asgiref==3.4.1
Django==4.0.1
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
